Make compiler able to compile JVM builtins as a part of stdlib
At the moment, JVM stdlib compilation is hacky: a subset of builtins
that do not produce bytecode is being compiled separately by
K2Metadata compiler and then used as a dependency when compiling the
rest of stdlib. We would like to compile them in a bunch with the rest
of sources, see KT-68154 for details.

To do so, we need some fixes in the compiler. Since the stdlib is being
compiled by bootstrap compiler, we need to merge compiler changes
and advance the bootstrap version before merging new build scripts
and adjusted stdlib sources.

Also, since the new compiler cannot work with the old stdlib build
scripts, the (part of) new behavior is disabled by default and enabled
by the `-Xcompile-builtins-as-part-of-stdlib` key.

No tests have been added, as stdlib compilation is itself an exhaustive
test.