Make compiler able to compile JVM builtins as a part of stdlib
At the moment, JVM stdlib compilation is hacky: a subset of builtins
that do not produce bytecode is being compiled separately by
K2Metadata compiler and then used as a dependency when compiling the
rest of stdlib. We would like to compile them in a bunch with the rest
of sources, see KT-68154 for details.

To do so, we need some fixes in the compiler. The most specificity of
compiling all the builtins in a bunch with other stdlib sources is
caused by so-called declaration enhancement, when the compiler maps
certain builtin declarations on JDK declarations:

1. Consider, e.g. `kotlin.Throwable` mapped on `java.lang.Throwable`,
and therefore gaining `getLocalizedMessage` declaration in its member
scope. The compiled did not create an IR declaration for it (and other
enhanced members) - however, it might be queried, at least as
overridden of its fake override in child classes. This class and some
other used to be external declarations, with different logic of IR
declaration creation, which used to hide the problem.

2. All the enhanced declarations had `isOverride == true`,
which does not make sense and breaks assertions of `FirOverrideChecker`.

3. Adding enhanced JDK members to serialized metadata of kotlin builtin
classes interferes badly with enhancing when compiling user code and
causes compilation errors on the call site. I have not investigated,
but I assume it is safe to simply filter them out, as they did not
present in old-way compiled stdlib, as well as special
`FakeHiddenInPreparationForNewJdk` members.

Apart from the enhancement, there are other special cases to handle:

1. @PlatformDependent declarations. When compiling stdlib, we don't
want to filter them out from the `JvmMappedScope` scope as usual, as
they should end up in the final serialized declaration list.

2. We can no longer rely on declaration origin to distinguish builtins
when compiling stdlib, as they are simply defined in sources now -
thus, we introduced special annotations. The annotations also are
useful to avoid codegen and lowering invocation for the declarations
which do not produce bytecode, and for deciding what should be
serialized in `.kotlin_builtins`.

Also, we no longer need ad-hoc @ActualizeByJvmBuiltinProvider-related
logic.

Since the stdlib is being compiled by bootstrap compiler, we need to
merge compiler changes and advance the bootstrap version before merging
new build scripts and adjusted stdlib sources.

Also, since the new compiler cannot work with the old stdlib build
scripts, the (part of) new behavior is disabled by default and enabled
by the `-Xcompile-builtins-as-part-of-stdlib` key. The old behavior will
not be necessary after merging new build scripts into master, so we will
drop the old logic after it. When bootstrap is advanced to the version
with new default logic, we will drop the flag itself as well.

No tests have been added, as stdlib compilation is itself an exhaustive
test.