[Wasm, IC] Fix invalid incremental signature for the different IC steps
The different IC steps can have a functions with default arguments when it loaded by IrLinker. IrLinker could load a function as implemented regular function and as FakeOverride when declaration is not needed to be loaded. So DefaultArgumentFunctionFactory creates default's overload based on (i.e. restrictedBy) different functions so two IC step can produce a different signature for the same functions. To eliminate this we have to put restrictions on the function which would be used on callsite.
(the bug is dependent from lowering file sequence, so we test it with file renaming)

Fixed #KT-72232